+++ examples/cvode/serial/cvDiurnal_kry_bp.c
@@ -294,7 +294,7 @@ static void PrintIntro(sunindextype mu, sunindextype ml)
   printf("2-species diurnal advection-diffusion problem, %d by %d mesh\n",
          MX, MY);
   printf("SPGMR solver; band preconditioner; mu = %ld, ml = %ld\n\n",
-         mu, ml);
+         (long int) mu, (long int) ml);
 
   return;
 }
@@ -387,9 +387,9 @@ static void PrintFinalStats(void *cvode_mem)
   check_flag(&flag, "CVBandPrecGetNumRhsEvals", 1);
 
   printf("\nFinal Statistics.. \n\n");
-  printf("lenrw   = %5ld     leniw   = %5ld\n", lenrw, leniw);
-  printf("lenrwls = %5ld     leniwls = %5ld\n", lenrwLS, leniwLS);
-  printf("lenrwbp = %5ld     leniwbp = %5ld\n", lenrwBP, leniwBP);
+  printf("lenrw   = %5ld     leniw   = %5ld\n", (long int) lenrw, (long int) leniw);
+  printf("lenrwls = %5ld     leniwls = %5ld\n", (long int) lenrwLS, (long int) leniwLS);
+  printf("lenrwbp = %5ld     leniwbp = %5ld\n", (long int) lenrwBP, (long int) leniwBP);
   printf("nst     = %5ld\n"                  , nst);
   printf("nfe     = %5ld     nfetot  = %5ld\n"  , nfe, nfe+nfeLS+nfeBP);
   printf("nfeLS   = %5ld     nfeBP   = %5ld\n"  , nfeLS, nfeBP);
